Flipping cat

+++ Amabili_Eilers_OpenGL_3/OpenGL_3_startpoint/Code/properties.cpp
@@ -1,6 +1,6 @@
 #include "properties.h"
 
-void setProperties(QVector3D location, QVector3D rotation, float speed, int meshSize, Properties &p) {
+void setProperties(QVector3D location, QVector3D rotation, double speed, int meshSize, Properties &p) {
     p.location = location;
     p.rotation = rotation;
     p.speed = speed;
